Improve code quality
Code quality does not adhere to SLAP principles.

Adherance to SLAP principles is needed to ensure common levels of abstractions
in methods of the Command and Task related classes.

Long methods are broken down into more methods, with each method handling a
small portion of the logic.

This is to ensure that the methods are not overly complex and adhere to SLAP
principles.